requirements.txt
streamlit
transformers
requests
python-dotenv
